requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic[email]>=2.10.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
resend>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
